LayoutTests/fast/dom/DOMURL/check-instanceof-domurl-functions-expected.txt

Test instanceof functions and properties of webkitURL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'createObjectURL' in webkitURL is true
PASS 'revokeObjectURL' in webkitURL is true
PASS url.createObjectURL is undefined.
PASS url.revokeObjectURL is undefined.
PASS window.webkitURL.createObjectURL is defined.
PASS window.webkitURL.revokeObjectURL is defined.
PASS successfullyParsed is true

TEST COMPLETE
